Fix customer total spend and amount type of Invoices table

+++ database/migrations/2020_11_24_172800_create_invoice_table.php
@@ -34,7 +34,7 @@ class CreateInvoiceTable extends Migration
             ->on('intakes')
             ->onUpdate('cascade');
 
-            $table->integer('amount');
+            $table->double('amount',10,2);
             $table->integer('promotion_amount')->nullable();
             $table->text('note')->nullable();
             $table->binary('signature')->nullable();
